# src/server/openapi-ai.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Priority App API",
    "version": "1.0.0",
    "description": "AI-optimized task management API with ICE scoring (Impact, Confidence, Ease), 4D decision framework (Do, Delegate, Delay, Delete), time tracking, and AI recommendations. Use this API to manage tasks, track time, get productivity analytics, and manage programmatic access via API keys.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/task-priority-app"
    },
    "x-ai-friendly": true,
    "x-supports-function-calling": true
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    },
    {
      "url": "https://your-app.workers.dev",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication. Use format: 'Bearer task_live_XXXXXXXX...' in Authorization header"
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "description": "A task with ICE scoring for prioritization",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique task identifier"
          },
          "name": {
            "type": "string",
            "description": "Task title or name",
            "example": "Complete project documentation"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes or description in plain text or markdown"
          },
          "impact": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Impact score (1-10): How much value this task provides when completed",
            "example": 8
          },
          "confidence": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Confidence score (1-10): How confident you are in your ability to complete this",
            "example": 7
          },
          "ease": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Ease score (1-10): How easy/quick this task is to complete",
            "example": 6
          },
          "type": {
            "type": "string",
            "enum": ["revenue", "growth", "operations", "strategic", "personal"],
            "description": "Task category: revenue (income-generating), growth (learning/skills), operations (day-to-day), strategic (long-term planning), personal (personal tasks)"
          },
          "timeBlock": {
            "type": "string",
            "enum": ["deep", "collaborative", "quick", "systematic"],
            "description": "When to do this: deep (focused deep work), collaborative (with others), quick (small tasks), systematic (routine processes)"
          },
          "estimatedTime": {
            "type": "integer",
            "description": "Estimated duration in minutes",
            "example": 120
          },
          "decision": {
            "type": "string",
            "enum": ["do", "delegate", "delay", "delete"],
            "description": "4D Framework decision: do (prioritize now), delegate (assign to others), delay (postpone), delete (eliminate)"
          },
          "status": {
            "type": "string",
            "enum": ["active", "completed", "archived"],
            "default": "active",
            "description": "Task status"
          },
          "actualTime": {
            "type": "integer",
            "default": 0,
            "description": "Total time spent on task in minutes (tracked via focus sessions)"
          },
          "isInFocus": {
            "type": "boolean",
            "description": "Whether task currently has an active time tracking session"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Hard deadline for task completion (ISO 8601 format)",
            "example": "2025-12-31T23:59:59Z"
          },
          "scheduledFor": {
            "type": "string",
            "enum": ["today", "this-week", "this-month", "someday"],
            "nullable": true,
            "description": "Time horizon for task completion"
          },
          "recurringPattern": {
            "type": "string",
            "enum": ["daily", "weekly", "monthly"],
            "nullable": true,
            "description": "Recurrence pattern for repeating tasks"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When task was created"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When task was completed"
          }
        },
        "required": ["name", "impact", "confidence", "ease", "type", "timeBlock", "estimatedTime", "decision"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "List all tasks",
        "description": "Retrieve all tasks for the authenticated user. Optionally filter by status (active, completed, archived). Returns tasks with ICE scores, time tracking data, and decision framework classification.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter tasks by status",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "archived"],
              "example": "active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTask",
        "summary": "Create a new task",
        "description": "Create a new task with ICE scoring (Impact, Confidence, Ease). The task will be prioritized using AI algorithms based on these scores and the 4D decision framework (Do, Delegate, Delay, Delete).",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "description": "Task details including required ICE scores and categorization",
          "content": {
            "application/json": {
                "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Validation error",
                  "details": {
                    "impact": "Must be between 1 and 10"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "operationId": "getTask",
        "summary": "Get a specific task",
        "description": "Retrieve detailed information about a specific task by its ID, including all ICE scores, time tracking data, and status.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateTask",
        "summary": "Update a task",
        "description": "Update specific fields of a task. You can update any combination of fields including ICE scores, status, name, or decision classification.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update (all optional, only include fields you want to change)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "completed", "archived"]
                  },
                  "impact": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "confidence": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "ease": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "decision": {
                    "type": "string",
                    "enum": ["do", "delegate", "delay", "delete"]
                  }
                },
                "example": {
                  "status": "completed",
                  "impact": 9
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "summary": "Delete a task",
        "description": "Permanently delete a task and all associated time tracking data. This action cannot be undone.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Task deleted successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/focus/start": {
      "post": {
        "operationId": "startFocusSession",
        "summary": "Start time tracking",
        "description": "Start a focus/time tracking session for a task. This automatically stops any other active focus sessions to ensure only one timer runs at a time. Use this when you begin working on a task.",
        "tags": ["Time Tracking"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID to start tracking",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Focus session started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/focus/end": {
      "post": {
        "operationId": "endFocusSession",
        "summary": "Stop time tracking",
        "description": "End a focus session and record the time spent. The duration is added to the task's total actualTime. Use this when you finish working on a task.",
        "tags": ["Time Tracking"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID to stop tracking",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Duration of the focus session",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["durationMinutes"],
                "properties": {
                  "durationMinutes": {
                    "type": "integer",
                    "description": "How many minutes were spent on this task",
                    "example": 25
                  }
                },
                "example": {
                  "durationMinutes": 25
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Focus session ended, time recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}